Normalize configurations to use Settings
There are many small changes here to normalize the configuration such
that it can be committed and have convenient overrides where needed. The
pattern of using .sample files meant that every environment must
generate copies, which leads to needless duplication and desynchronization,
making it difficult to manage multiple instances and changes over time.

The config gem in use allows for local overrides of specific settings,
rather than entire files. This means that generic files that refer to
Settings can be committed and an instance may customize only what it
needs; for example, the Solr core name.

There is also some reorganization and renaming of the settings and how
they are applied within the application, to improve consistency. For
example, the settings that are applied in the Hyrax initializer are
nested under a hyrax key. There are some notes of where further work
should be done.

Where possible, the YAML files use aliases to reduce the duplication of
key expressions. This makes files like fedora.yml very generic, but
there are comments that specific values should not be placed there and
references to the other files where they should be placed. The master
settings.yml file is commented to describe the sections and the settings
within them.

Re: AEIM-1906