Map settings into database.yml
This change makes database.yml like the other stock Rails files in that
the values from Settings are mapped in. This uses a different approach
where the keys are copies wholesale, rather than explicitly, because it
is very hard to predict which adapters could need values for which keys.

It retrieves the settings and dumps them as YAML to be inlined under a
hash with an alias. It's a bit of an ugly expression, but works well,
and should be robust in that it's round tripping through the same
parser. This could easily be turned into a library function to get it
out of the YAML file itself.

Re: AEIM-1906